+++ src/runtime/components/UniversalViewer.vue
@@ -1,0 +1,20 @@
+<template>
+  <LazyGLottieAnimation
+      v-if="type === 'lottie'"
+      ref="lottieRef"
+      v-bind="$attrs"
+  />
+  <NuxtImg
+      v-else
+      v-bind="$attrs"
+  />
+</template>
+
+<script setup>
+const props = defineProps({
+  type: {
+    type: String,
+    required: true, // "image" | "lottie"
+  }
+})
+</script>