+++ src/runtime/components/media/CreateMedia.vue
@@ -47,9 +47,9 @@
 </template>
 
 <script setup>
-import { useI18n, ref, useRoute, navigateTo, useLocalePath, watch, useCookie, useFetch } from '#imports'
+import { useI18n, ref, useRoute, navigateTo, useLocalePath, watch, useFetch } from '#imports'
 
-import { acceptedFileTypes, mediaHeader, showToast } from './medias'
+import {acceptedFileTypes, isLottieAnimation, mediaHeader, showToast} from './medias'
 
 const { t } = useI18n()
 
@@ -163,6 +163,23 @@ async function onFileSelected(e) {
   data.append('type', fileData.type.includes('image') ? 'image' : 'document')
   data.append('private_status', 'public')
   data.append('locked_status', 'unlocked')
+
+  function readFileAsText(file) {
+    return new Promise((resolve, reject) => {
+      const reader = new FileReader()
+      reader.onload = e => resolve(e.target.result)
+      reader.onerror = reject
+      reader.readAsText(file)
+    })
+  }
+
+  try {
+    const fileText = await readFileAsText(fileData)
+    const isLottie = isLottieAnimation(JSON.parse(fileText))
+    if (isLottie) {
+      data.append('metadata[type]', 'lottie')
+    }
+  } catch {}
 
   try {
     const res = await props.requestPreSent('POST', mediaByIdUri.value, data)